external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/glsl/reserved/webgl_variable.vert-expected.txt

CONSOLE MESSAGE: line 24: *** Error compiling shader '[object WebGLShader]':ERROR: 0:3: 'webgl_' : reserved built-in name 

use of reserved webgl_ prefix as variable should fail

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

test: use of reserved webgl_ prefix as variable should fail
test vertex shader
default fragment shaderPASS use of reserved webgl_ prefix as variable should fail
PASS successfullyParsed is true

TEST COMPLETE
